Spawn goroutines from a loop to print words
A simple program demonstrating that goroutines can also be spawned
from a loop. The order of the resulting printed text varies
heavily from runtime to runtime; but in my tests the delay of 500ms
at the end of main() was big enough that all the 10 items were
printed at every run.